data-management-backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
databases==0.8.0
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
PyYAML==6.0.1
setuptools==68.2.2
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.25.0
wheel==0.41.2
